Fix warnings emitted by `clang` (#10)
* Remove extra argument to function.

* Remove trivial comparison.

An array is never NULL.

* Explicitly cast a return value from a float to an int.

* Fixed signedness in our lzw implementation.

* Don't call a no-op function (with the wrong number of arguments).

+++ winbolo/src/bolo/bases.c
@@ -1618,7 +1618,7 @@ int basesHalfTickCalulator(int typeSelector) {
 			return (int) floor(tempShell);
 		}
 	  } else {
-	    return BASE_REFUEL_SHELLS;
+	    return (int) BASE_REFUEL_SHELLS;
 	  }
 	  break;
 	case BASES_HALFTICK_TYPE_MINE: 
@@ -1632,7 +1632,7 @@ int basesHalfTickCalulator(int typeSelector) {
 			return (int) floor(tempMine);
 		}
 	  } else {
-	    return BASE_REFUEL_MINES;
+	    return (int) BASE_REFUEL_MINES;
 	  }
 	  break;
 	case BASES_HALFTICK_TYPE_ARMOUR:
